src/ecr-image-list/requirements.txt
aws-cdk-lib>=2.0.0
aws-cdk.aws-lambda-python-alpha
constructs>=10.0.0
boto3
requests
